% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_association.R
\name{jitter.columns}
\alias{jitter.columns}
\title{Adds jitter to the input dataframe to break ties.}
\usage{
jitter.columns(data, factor = 0.1)
}
\arguments{
\item{data}{Input dataframe of data}

\item{factor}{A float to scale the jittering by (if greater than 2.5, there is a risk of quantile crossings)}
}
\value{
Jittered version of your input dataframe
}
\description{
Goes column by column in the dataframe and adds jitter from the distribution
}
\details{
Unif(-a, a)

where a = factor * d /5 (and d is the minimum distance between any two points in a vector).
}
