% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_association.R
\name{marginal.test}
\alias{marginal.test}
\title{Returns the marginal quantile association for all pairs of variables in a dataframe.}
\usage{
marginal.test(data, tau)
}
\arguments{
\item{data}{Input dataframe of data}

\item{tau}{A particular quantile level (0 to 1, not inclusive)}
}
\value{
An n by n matrix (where n is the number of columns in data) that contains the marginal relationships of each pair of columns
}
\description{
Runs a double for loop across all columns and calculates the quantile association
statistic for all possible pairs. Depending on if the quantile level is above or
below the median, the statistic will consider if the residuals are jointly above or
jointly below their respective marginal regression lines.
}
