% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_association.R
\name{pairwise.test}
\alias{pairwise.test}
\title{Returns the  quantile association for all pairs of variables in a dataframe.}
\usage{
pairwise.test(data, tau, weights = "marginal")
}
\arguments{
\item{data}{Input dataframe of data}

\item{tau}{A particular quantile level (0 to 1, not inclusive)}

\item{type}{If weights of the regression should be marginal or conditional on all other variables}
}
\value{
An n by n matrix (where n is the number of columns in data) that contains the marginal relationships of each pair of columns
}
\description{
Runs a double for loop across all columns and calculates the quantile association
statistic for all possible pairs. Depending on if the quantile level is above or
below the median, the statistic will consider if the residuals are jointly above or
jointly below their respective marginal regression lines.
}
\details{
This test has two types where it considers marginal associations of variables or
considers the conditional case where each quantile regression is conditional w.r.t
all other variables in the dataframe.
}
