% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.marginal.table}
\alias{plot.marginal.table}
\title{Calculates the table of marginal relationships at a particular quantile level.}
\usage{
\method{plot}{marginal.table}(df, tau, weights = "marginal", split = FALSE)
}
\arguments{
\item{tau}{A particular quantile level (0 to 1, not inclusive)}

\item{weights}{Weights will either be "marginal" (marginal association)
or "conditional" (association conditioned on everything) - default is
"marginal" (otherwise they will not be used)}

\item{split}{If you should split the data for more efficient estimates (fits twice as many models)}

\item{data}{A data frame of data with unique column/rownames}
}
\value{
Nothing
}
\description{
It assumes the input dataframe has at least two columns (columns are variables) and that
the quantile level is between 0 and 1 (non-inclusive). Additionally it assumes all
column names are unique.
}
\details{
This function uses calculations from the quantile concordance statistic
in pairwise.test and plots it in a nicer table format.

If weights is equal to "cor" the pariwise Pearson correlation table is
returned.
}
\examples{
plot.marginal.table(df, tau=0.5, weights="marginal")
}
