% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{qgm.igraph.plot}
\alias{qgm.igraph.plot}
\title{Converts the pcalgo graph object into an igraph object and plots it.}
\usage{
qgm.igraph.plot(
  g,
  weights = TRUE,
  layout = igraph::layout_in_circle,
  vertex.size = 50,
  vertex.color = "lightblue",
  ...
)
}
\arguments{
\item{g}{A graph object of class graph}

\item{weights}{A list of weights for edges in the graph}

\item{layout}{A layout to organize the graph with}

\item{vertex.size}{A node size for the graph}

\item{vertex.color}{Specify the color of nodes in the graph}
}
\value{
Nothing
}
\description{
Constructs igraph object from graphnel type and adds weights (if passed in).
Default layout is in a circle so graphs are easily comparable.
}
