% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_association.R
\name{quantile.ztest}
\alias{quantile.ztest}
\title{Performs a hypothesis test of two variables given a conditioning set and returns
a p-value on if they are independent (p > 0.05) or not (p <= 0.05).}
\usage{
\method{quantile}{ztest}(x, y, S, suffStat)
}
\arguments{
\item{x}{Index of a column}

\item{y}{Index of a column (not equal to x)}

\item{S}{Conditioning set to be used to determine if conditional independence exists (can be empty set)}

\item{suffStat}{The dataframe of data (there is no sufficient statistic for this calculation)}
}
\value{
pvalue corresponding to if the quantile level
}
\description{
Uses a hypothesis test that considers the proportion of times two variables
have concordant (conditional) quantile regression residuals. Counts the
amount of times the residual pairs are concordant (jointly above or below their
respective conditional quantile regression lines) and normalizes that concordant
proportion. Normalization is based on what the proportions would be
under the null hypothesis. From the normalized variable, p values of the
relationship between columns x and y can be returned.
}
